requirements.txt
numpy==2.2.6
pillow==11.2.1
pygame==2.6.1
RapidFuzz==3.13.0
Wand==0.6.13
